aula2/requirements.txt
flask
sqlalchemy
python-dotenv
psycopg2-binary

